make Server code cleaner.
